refactor(credital management tests): load default configuration
