+++ handlers/iam_handler.go
@@ -1,4 +1,4 @@
-package aws
+package handlers
 
 import (
 	"net/http"
